requirements.txt
langchain-core
openai
python_dotenv